+++ packages/grid/src/features/standard-table/features/checkboxes/StandardTableRowCheckbox.tsx
@@ -3,7 +3,10 @@ import { Checkbox, CheckboxProps } from "@stenajs-webui/forms";
 import * as React from "react";
 import { useGridCell } from "../../../grid-cell/hooks/UseGridCell";
 import { useTotalNumColumnsForRows } from "../../context/GroupConfigsForRowsContext";
-import { useStandardTableId } from "../../hooks/UseStandardTableConfig";
+import {
+  useStandardTableConfig,
+  useStandardTableId,
+} from "../../hooks/UseStandardTableConfig";
 
 interface Props extends Pick<CheckboxProps, "value" | "onValueChange"> {
   colIndex: number;
@@ -24,6 +27,11 @@ export const StandardTableRowCheckbox: React.FC<Props> = React.memo(
     const totalNumColumns = useTotalNumColumnsForRows();
 
     const tableId = useStandardTableId();
+    const {
+      zIndex,
+      stickyHeaderAndRowCheckbox,
+      stickyRowCheckboxRightShadow,
+    } = useStandardTableConfig();
     const gridCell = useGridCell<boolean>(Boolean(value), {
       colIndex,
       numCols: totalNumColumns,
@@ -40,6 +48,15 @@ export const StandardTableRowCheckbox: React.FC<Props> = React.memo(
         width={"45px"}
         minWidth={"45px"}
         indent
+        background={stickyHeaderAndRowCheckbox ? "inherit" : undefined}
+        position={stickyHeaderAndRowCheckbox ? "sticky" : undefined}
+        left={stickyHeaderAndRowCheckbox ? "0px" : undefined}
+        shadow={
+          stickyHeaderAndRowCheckbox && stickyRowCheckboxRightShadow
+            ? "var(--swui-sticky-column-shadow-right)"
+            : undefined
+        }
+        zIndex={zIndex}
       >
         <Checkbox
           size={"small"}